services/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
kafka-python
minio
pydantic
prometheus-client
boto3
python-dotenv
requests
clickhouse-driver==0.2.7
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

